[ADD] implement jsend specification [ADD] use jsend specification to return json objects for all ajax routes [IMP] robustness of api with better type and value checks

+++ ip_web_addons/account/account.py
@@ -4,7 +4,9 @@ from openerp.addons.web import http
 from openerp.addons.web.http import request
 from openerp.addons.website.models import website
 from openerp.tools.translate import _
- 
+
+from .. import jsend, tools
+
 class IpMyAccount(http.Controller):
 
     @http.route(['/account/', '/account/<page>'], type='http', auth="public", multilang=True, website=True)
@@ -80,30 +82,56 @@ class IpMyAccount(http.Controller):
 
         return request.website.render("ip_web_addons.account", vals)
     
+    @jsend.jsend_error_catcher
     @http.route(['/account/auto-ship/update/<int:auto_ship_id>'], type='http', auth="public", multilang=True, website=True)
     def update_auto_ship(self, auto_ship_id, interval, end_date, **post):
         """ Update an auto ship's interval and end date """
-        assert auto_ship_id and interval and end_date, "All variables are required to be truthy"
+        if not isinstance(auto_ship_id, (int, float)):
+            raise jsend.JsendTypeError("auto_ship_id", "Auto ship ID should be a number")
+        if not auto_ship_id:
+            raise jsend.JsendValueError("auto_ship_id", "Auto ship ID must be greater than 0")
+        
+        if not tools.isnumeric(interval):
+            raise jsend.JsendTypeError("interval", "Interval should be a number")
+        if not interval:
+            raise jsend.JsendValueError("interval", "Interval should be greater than 0") 
+         
+        if not isinstance(end_date, (str, unicode)):
+            raise jsend.JsendTypeError("end_date", "End Date was invalid")
+        if not end_date:
+            raise jsend.JsendValueError("end_date", "End Date is required")  
+        
         cr, uid, pool = request.cr, request.uid, request.registry
         # TODO: do we need to check permission?
         pool['ip.auto_ship'].write(cr, uid, auto_ship_id, {'interval': interval, 'end_date': end_date})
-        return 'Your changes have been saved'
     
-    @http.route(['/account/auto-ship/delete/<int:auto_ship_id>'], type='http', auth="public", multilang=True, website=True)
+    @jsend.jsend_error_catcher
+    @http.route(['/account/auto-ship/delete/auto_ship_id'], type='http', auth="public", multilang=True, website=True)
     def delete_auto_ship(self, auto_ship_id, **post):
         """ delete an auto ship """
-        assert auto_ship_id and isinstance(auto_ship_id, (int, long)), "Invalid auto ship ID"
+        if not tools.isnumeric(auto_ship_id):
+            raise jsend.JsendTypeError("auto_ship_id", "Auto ship ID should be a number")
+        if not auto_ship_id:
+            raise jsend.JsendValueError("auto_ship_id", "Auto ship ID should be greater than 0")
+        
         cr, uid, context, pool = request.cr, request.uid, request.context, request.registry
         # TODO: do we need to check permission?
         pool['ip.auto_ship'].unlink(cr, uid, auto_ship_id, context=context)
-        return 'true'
-
-    @http.route(['/account/number-remaining/<int:interval>/<end_date>'], type='http', auth="public", multilang=True, website=True)
+        
+    @jsend.jsend_error_catcher
+    @http.route(['/account/number-remaining/<interval>/<end_date>'], type='http', auth="public", multilang=True, website=True)
     def number_remaining(self, interval, end_date, **post):
         """ Return the number of auto shipments remaining based on end_date and interval """
-        if not interval or not end_date:
-            return ''
+        if not tools.isnumeric(interval):
+            raise jsend.JsendTypeError("interval", "Interval should be a number")
+        if not interval:
+            raise jsend.JsendValueError("interval", "Interval should be greater than 0")  
+        if not isinstance(end_date, (str, unicode)):
+            raise jsend.JsendTypeError("end_date", "End Date was invalid")
+        if not end_date:
+            raise jsend.JsendValueError("end_date", "End Date is required")  
         try:
-            return str(request.registry['ip.auto_ship']._calculate_number_remaining(float(interval), end_date))
-        except:
-            return ''
+            nr = str(request.registry['ip.auto_ship']._calculate_number_remaining(float(interval), end_date))
+            return jsend.jsend_success({'number_remaining': nr})
+        except ValueError as e:
+            raise jsend.JsendValueError("end_date", "End Date must be a valid date in YYYY-MM-DD format")
